+++ WebStore/Controllers/CatalogController.cs
@@ -12,30 +12,53 @@ namespace WebStore.Controllers
 
         public CatalogController(IProductData ProductData) => _ProductData = ProductData;
 
-        public IActionResult Index(int? BrandId, int? SectionId)
+        public IActionResult Index(int? brandId, int? sectionId)
         {
             var filter = new ProductFilter
             {
-                BrandId = BrandId,
-                SectionId = SectionId,
+                BrandId = brandId,
+                SectionId = sectionId,
             };
 
             var products = _ProductData.GetProducts(filter);
 
-            return View(new CatalogViewModel
+            var model = new CatalogViewModel()
             {
-                SectionId = SectionId,
-                BrandId = BrandId,
-                Products = products
-                   .OrderBy(p => p.Order)
-                   .Select(p => new ProductViewModel
-                   {
-                       id = p.id,
-                       Name = p.Name,
-                       Price = p.Price,
-                       ImageUrl = p.ImageUrl,
-                   })
-            });
+                BrandId = brandId,
+                SectionId = sectionId,
+                Products = products.Select(p => new ProductViewModel()
+                {
+                    id = p.id,
+                    ImageUrl = p.ImageUrl,
+                    Name = p.Name,
+                    Order = p.Order,
+                    Price = p.Price
+                }).OrderBy(p => p.Order).ToList()
+            };
+
+            return View(model);
         }
+
+        public IActionResult Shop(int? sectionId, int? brandId)
+        {
+            var products = _ProductData.GetProducts(new ProductFilter { BrandId = brandId, SectionId = sectionId });
+
+            var model = new CatalogViewModel()
+            {
+                BrandId = brandId,
+                SectionId = sectionId,
+                Products = products.Select(p => new ProductViewModel()
+                {
+                    id = p.id,
+                    ImageUrl = p.ImageUrl,
+                    Name = p.Name,
+                    Order = p.Order,
+                    Price = p.Price
+                }).OrderBy(p => p.Order).ToList()
+            };
+
+            return View(model);
+        }
+
     }
 }
