+++ WebStore/ViewModels/ProductViewModel.cs
@@ -1,13 +1,26 @@
-﻿namespace WebStore.ViewModels
+﻿using System.Collections.Generic;
+
+namespace WebStore.ViewModels
 {
     public class ProductViewModel
     {
         public int id { get; set; }
+
+        public int Order { get; set; }
 
         public string Name { get; set; }
 
         public string ImageUrl { get; set; }
 
         public decimal Price { get; set; }
+    }
+
+    public class CatalogViewModel
+    {
+        public IEnumerable<ProductViewModel> Products { get; set; }
+
+        public int? SectionId { get; set; }
+
+        public int? BrandId { get; set; }
     }
 }
