+++ WebStore/Views/Catalog/Shop.cshtml
@@ -3,7 +3,7 @@
     ViewData["Title"] = "Catalog";
 }
 
-@await Html.PartialAsync("Partial/_Advertisement")
+<partial name="Partial/_Advertisement" />
 
 <section>
     <div class="row">
