+++ WebStore/Views/Shared/Partial/_MainMenu.cshtml
@@ -3,7 +3,7 @@
     <li class="dropdown">
         <a asp-controller="Home" asp-action="Shop">Shop<i class="fa fa-angle-down"></i></a>
         <ul role="menu" class="sub-menu">
-            <li><a asp-controller="Home" asp-action="ProductDetails">Products</a></li>
+            <li><a asp-controller="Catalog" asp-action="Shop">Products</a></li>
             <li><a asp-controller="Home" asp-action="ProductDetails">Product Details</a></li>
             <li><a asp-controller="Home" asp-action="Checkout">Checkout</a></li>
             <li><a asp-controller="Home" asp-action="Cart">Cart</a></li>
